Modify the logic code inside play funtion to
work properly. I was unable to identify the problem myself so took help I was basically also giving the logical condition when they were not needed.the code is working and can show the score of the computer or human but cannot display the return string of win instead only prints the "Tie".